Extract Manifest() into manifest.go for maintainability
Split shopify.go (516 lines) into shopify.go (205 lines) for connector
logic and manifest.go (318 lines) for action schemas, credentials, and
templates. This keeps each file focused on a single concern and makes it
easier to add new actions without scrolling past the HTTP client code.

No functional changes — just a file split. All tests pass with -race.
